\name{Silly-class}
\Rdversion{1.1}
\docType{class}
\alias{Silly-class}

\title{S4 classes can be documented as well}
\description{The Silly class does nothing much either}
\section{Objects from the Class}{but you can put what you like in documentation details}
\section{Slots}{
  \describe{
    \item{\code{forwards}:}{(\code{function}) forward operation}
    \item{\code{reverse}:}{(\code{function}) how to go backward}
    \item{\code{crashes}:}{(\code{integer}) how many crashes}
  }
}
\section{Methods}{
No methods defined with class "Silly" in the signature.
}
\references{}
\author{Keith Ponting <k.ponting@aurix.com>}
\note{}



\seealso{}
\examples{}
