\name{silly-package}
\alias{silly-package}
\docType{package}
\title{Exhaustive example of inlinedocs extended tags}
\description{Attempt to exercise many variants of the extended ##<< tags for inline docs.}
\details{
\tabular{ll}{Package: \tab silly\cr
Title: \tab Exhaustive example of inlinedocs extended tags\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-11-12\cr
Author: \tab Keith Ponting\cr
Maintainer: \tab Keith Ponting <k.ponting@aurix.com>\cr
URL: \tab http://inlinedocs.r-forge.r-project.org\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr}
}
\author{Keith Ponting <k.ponting@aurix.com>}
\references{
}
\keyword{ package }
\seealso{
}
\examples{}
